Added in player death and respawn location.

+++ src/main/java/net/hollowbit/archipeloserver/entity/living/player/PlayerStatsManager.java
@@ -152,7 +152,7 @@ public class PlayerStatsManager implements EventHandler {
 		return speedMultiplier;
 	}
 	
-	public int getHealth() {
+	public float getHealth() {
 		return player.getHealth();
 	}
